switched to old impl. of decoloring, plus integration tests using maven-invoker-plugin
