+++ src/it/it-decoloring/invoker.properties
@@ -1,0 +1,1 @@
+invoker.goals = clean test
