+++ src/it/settings.xml
@@ -1,5 +1,6 @@
-/**
- * Copyright (c) 2012-2015, jcabi.com
+<?xml version="1.0"?>
+<!--
+ * Copyright (c) 2013-2015, jcabi.com
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -26,22 +27,38 @@
  * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
  * OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package com.jcabi.log;
-
-/**
- * Used to decide whether log text
- * produced by {@link MulticolorLayout} should be colored.
- * We cannot read the com.jcabi.log.coloring directly in
- * {@link MulticolorLayout} because that is not testable (tests
- * are run by multiple threads at once).
- * @author Mihai Andronache (amihaiemil@gmail.com)
- * @version $Id$
- */
-public interface TextDecolorant {
-    /**
-     * Decide whether text coloring is enabled or not.
-     * @return True if it is enabled or false otherwise.
-     */
-    boolean isColoringEnabled();
-}
+ -->
+<settings>
+    <profiles>
+        <profile>
+            <id>it-repo</id>
+            <activation>
+                <activeByDefault>true</activeByDefault>
+            </activation>
+            <repositories>
+                <repository>
+                    <id>local.central</id>
+                    <url>@localRepositoryUrl@</url>
+                    <releases>
+                        <enabled>true</enabled>
+                    </releases>
+                    <snapshots>
+                        <enabled>true</enabled>
+                    </snapshots>
+                </repository>
+            </repositories>
+            <pluginRepositories>
+                <pluginRepository>
+                    <id>local.central</id>
+                    <url>@localRepositoryUrl@</url>
+                    <releases>
+                        <enabled>true</enabled>
+                    </releases>
+                    <snapshots>
+                        <enabled>true</enabled>
+                    </snapshots>
+                </pluginRepository>
+            </pluginRepositories>
+        </profile>
+    </profiles>
+</settings>
